we dont really need such a large variance in size cast

+++ inc/util.hh
@@ -16,19 +16,19 @@ namespace aquinas
 {
     namespace util
     {
-        constexpr size_t KB_TO_BYTES = 1024;
-        constexpr size_t MB_TO_BYTES = 1024 * 1024;
+        constexpr unsigned KB_TO_BYTES = 1024;
+        constexpr unsigned MB_TO_BYTES = 1024 * 1024;
 
         // C++ REWRITE OF THE COMMON UTILIIES I HAVE PRE-SUPPOSED
         // FOR MEMORY OPERATIONS WHICH REQUIRE A PROPER FORMATTING OF SIZE
         
-        inline std::string FORMAT_UNIT(size_t SIZE)
+        inline std::string FORMAT_UNIT(unsigned SIZE)
         {
             return SIZE >= MB_TO_BYTES ? "MB" : 
                     SIZE >= KB_TO_BYTES ? "KB" : "B";
         }
 
-        constexpr auto FORMAT_SIZE(size_t SIZE)
+        constexpr auto FORMAT_SIZE(unsigned SIZE)
         {
             return SIZE >= MB_TO_BYTES ? SIZE / MB_TO_BYTES :
            SIZE >= KB_TO_BYTES ? SIZE / KB_TO_BYTES : SIZE;
